astroloji2/pyproject.toml
[project]
name = "astroloji-ai"
version = "0.1.0"
description = "Production-grade orchestration layer for Astroloji AI horoscope generation."
authors = [{ name = "Astroloji AI Team" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0,<0.112",
  "uvicorn[standard]>=0.30.0,<0.31",
  "pydantic>=2.7.0,<3.0",
  "python-dotenv>=1.0.1,<1.1",
  "langchain>=0.1.20,<0.2",
  "langchain-openai>=0.1.6,<0.2",
  "openai>=1.30.0,<2.0",
  "chromadb>=0.4.24,<0.5",
  "prometheus-client>=0.20.0,<0.21",
  "structlog>=24.1.0,<25.0",
  "cachetools>=5.3.3,<6.0",
  "skyfield>=1.49,<2.0",
  "numpy>=1.26.0,<2.0",
  "tenacity>=8.2.3,<9.0",
  "httpx>=0.27.0,<0.28"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0",
  "pytest-asyncio>=0.23.6,<0.24",
  "respx>=0.20.2,<0.21",
  "ruff>=0.4.3,<0.5",
  "mypy>=1.10.0,<1.11",
  "types-python-dateutil>=2.9.0.20240316",
  "types-requests>=2.31.0.20240406"
]

[project.scripts]
astroloji-api = "astroloji_ai.api.main:run"
kb-ingest = "astroloji_ai.rag.ingest:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
